migrate: fix error if gallery dir is missing

+++ tools/migrate.js
@@ -234,24 +234,26 @@ let migration = {
     do: async () => {
       let processed = 0
       console.info('Moving gallery files to db')
-      for (let fname of fs.readdirSync('public/dist/gallery')) {
-        let data = Buffer.from(fs.readFileSync('public/dist/gallery/' + fname)).toString('base64')
-        let type = null
+      if (fs.existsSync('public/dist/gallery')) {
+        for (let fname of fs.readdirSync('public/dist/gallery')) {
+          let data = Buffer.from(fs.readFileSync('public/dist/gallery/' + fname)).toString('base64')
+          let type = null
 
-        if (fname.endsWith('ogg')) type = 'audio/ogg'
-        if (fname.endsWith('mp3')) type = 'audio/mp3'
-        if (fname.endsWith('mp4')) type = 'video/mp4'
-        if (fname.endsWith('jpg')) type = 'image/jpg'
-        if (fname.endsWith('png')) type = 'image/png'
-        if (fname.endsWith('gif')) type = 'image/gif'
+          if (fname.endsWith('ogg')) type = 'audio/ogg'
+          if (fname.endsWith('mp3')) type = 'audio/mp3'
+          if (fname.endsWith('mp4')) type = 'video/mp4'
+          if (fname.endsWith('jpg')) type = 'image/jpg'
+          if (fname.endsWith('png')) type = 'image/png'
+          if (fname.endsWith('gif')) type = 'image/gif'
 
-        if (type) {
-          data = 'data:' + type + ';base64,' + data
-          await global.db.engine.insert('overlays.gallery', { type, data })
+          if (type) {
+            data = 'data:' + type + ';base64,' + data
+            await global.db.engine.insert('overlays.gallery', { type, data })
+          }
+
+          fs.unlinkSync('public/dist/gallery/' + fname)
+          processed++
         }
-
-        fs.unlinkSync('public/dist/gallery/' + fname)
-        processed++
       }
       console.info(` => ${processed} processed`)
     }
